refactor: move palette to new `src/palette` directory
