+++ src/palette/palette.c
